# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VetCare Pro API - Logiciel vétérinaire
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /owners:
    get:
      operationId: listOwners
      tags: [owners]
      summary: Liste des propriétaires
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Liste des propriétaires
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Owner"
    post:
      operationId: createOwner
      tags: [owners]
      summary: Créer un propriétaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOwnerBody"
      responses:
        "201":
          description: Propriétaire créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
  /owners/{id}:
    get:
      operationId: getOwner
      tags: [owners]
      summary: Détails d'un propriétaire
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Propriétaire trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
        "404":
          description: Non trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOwner
      tags: [owners]
      summary: Mettre à jour un propriétaire
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOwnerBody"
      responses:
        "200":
          description: Propriétaire mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
    delete:
      operationId: deleteOwner
      tags: [owners]
      summary: Supprimer un propriétaire
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Supprimé
...
  /consultations/{id}/facture:
    post:
      operationId: generateFacture
      tags: [consultations]
      summary: Générer la facture d'une consultation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Facture générée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facture"
...
  /factures:
    get:
      operationId: listFactures
      tags: [factures]
      summary: Liste des factures
      parameters:
        - name: statut
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Liste des factures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FactureWithDetails"
  /factures/{id}:
    get:
      operationId: getFacture
      tags: [factures]
      summary: Détails d'une facture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Facture trouvée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FactureWithDetails"
    patch:
      operationId: updateFactureStatut
      tags: [factures]
      summary: Mettre à jour le statut d'une facture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFactureStatutBody"
      responses:
        "200":
          description: Facture mise à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facture"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Owner:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
        prenom:
          type: string
        email:
          type: ["string", "null"]
        telephone:
          type: string
        adresse:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - nom
        - prenom
        - telephone
        - createdAt
    CreateOwnerBody:
      type: object
      properties:
        nom:
          type: string
        prenom:
          type: string
        email:
          type: ["string", "null"]
        telephone:
          type: string
        adresse:
          type: ["string", "null"]
      required:
        - nom
        - prenom
        - telephone
    Patient:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
        espece:
          type: string
        race:
          type: ["string", "null"]
        sexe:
          type: string
        dateNaissance:
          type: ["string", "null"]
        poids:
          type: ["number", "null"]
        couleur:
          type: ["string", "null"]
        sterilise:
          type: boolean
        ownerId:
          type: integer
        antecedents:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - nom
        - espece
        - sexe
        - sterilise
        - ownerId
        - createdAt
    PatientWithOwner:
      allOf:
        - $ref: "#/components/schemas/Patient"
        - type: object
          properties:
            owner:
              $ref: "#/components/schemas/Owner"
...
    Consultation:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        veterinaire:
          type: string
        date:
          type: string
        statut:
          type: string
        motif:
          type: ["string", "null"]
        anamnese:
          type: ["string", "null"]
        examenClinique:
          type: ["string", "null"]
        examensComplementaires:
          type: ["string", "null"]
        diagnostic:
          type: ["string", "null"]
        diagnosticIA:
          type: ["string", "null"]
        ordonnance:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        poids:
          type: ["number", "null"]
        temperature:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - patientId
        - veterinaire
        - date
        - statut
        - createdAt
    ConsultationWithPatient:
      allOf:
        - $ref: "#/components/schemas/Consultation"
        - type: object
          properties:
            patient:
              $ref: "#/components/schemas/PatientWithOwner"
...
    Facture:
      type: object
      properties:
        id:
          type: integer
        consultationId:
          type: integer
        numero:
          type: string
        montantHT:
          type: number
        tva:
          type: number
        montantTTC:
          type: number
        statut:
          type: string
        dateEmission:
          type: string
        datePaiement:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - consultationId
        - numero
        - montantHT
        - tva
        - montantTTC
        - statut
        - dateEmission
        - createdAt
    FactureWithDetails:
      allOf:
        - $ref: "#/components/schemas/Facture"
        - type: object
          properties:
            consultation:
              $ref: "#/components/schemas/ConsultationWithPatient"
    UpdateFactureStatutBody:
      type: object
      properties:
        statut:
          type: string
        datePaiement:
          type: ["string", "null"]
      required:
        - statut
...