# lib/api-zod/src/generated/types/factureWithDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VetCare Pro API - Logiciel vétérinaire
 * OpenAPI spec version: 0.1.0
 */
import type { ConsultationWithPatient } from "./consultationWithPatient";
import type { Facture } from "./facture";

export type FactureWithDetails = Facture & {
  consultation?: ConsultationWithPatient;
};